Model Update
This update contains some major revisions and small implementations.

First every with statement has been removed from the methods within the class due to unpractical behavior in certain circumstances. It is now advised for the Controller to cast the with statement themselves whenever it needs to work with the database.

Additionally some methods got extra features. Like the dumpTable method, which got a reload function aswell as the getTable method, where now the the column can be selected.

Another small feature is the dumpDatabase method which essentially just deletes the whole .db file.

Also this commit contains some changes in the repo. They don't affect the code and are only meant to make the repo a bit clearer.